+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## PENDING
+
+### Bug Fixes
+
+- [#422](https://github.com/okta/okta-auth-js/pull/422) Fix bad accessToken storage key in signOut method
+
 ## 3,2.1
 
 ### Bug Fixes